added signing aws requests
